firmware: drivers: move motor driver classes to drivers folder
